requirements.txt
flask
groq
python-dotenv
gtts
gunicorn